restaurant and product fixed

+++ src/main/java/mimsoft/io/lemenu/product/ProductServiceImpl.java
@@ -2,12 +2,14 @@ package mimsoft.io.lemenu.product;
 
 import jakarta.transaction.Transactional;
 import lombok.SneakyThrows;
+import mimsoft.io.lemenu.content.TextModel;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class ProductServiceImpl implements ProductService {
@@ -22,37 +24,80 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     @Transactional
-    public boolean save(Product product) {
-        productRepository.save(product);
+    public boolean save(ProductDto productDto) {
+        productRepository.save(fromDto(productDto));
         return true;
     }
 
     @Override
-    public List<Product> getAll() {
-        return productRepository.findAll();
+    public List<ProductDto> getAll() {
+        return productRepository.findAll().stream()
+                .map(this::toDto)
+                .collect(Collectors.toList());
     }
 
 
     @Override
-    public Optional<Product> findById(Long id) {
-        return productRepository.findById(id);
+    public ProductDto findById(Long id) {
+        Optional<Product> productOptional = productRepository.findById(id);
+        if (productOptional.isPresent()) {
+            Product product = productOptional.get();
+            return toDto(product);
+        }
+        else
+            return null;
     }
 
     @SneakyThrows
     @Override
     @Transactional
-    public boolean update(Product product) {
-        productRepository.findById(product.getId())
-                .orElseThrow(() -> new RuntimeException("Product not found"));
-        productRepository.save(product);
-        return true;
+    public boolean update(ProductDto productDto) {
+        Optional<Product> productOptional = productRepository.findById(productDto.getId());
+        if (productOptional.isPresent()) {
+            productRepository.save(fromDto(productDto));
+            return true;
+        }
+        else
+            return false;
     }
 
     @Override
     public boolean delete(Long id) {
-        Product product = productRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Product not found"));
-        productRepository.delete(product);
+        Optional<Product> productOptional = productRepository.findById(id);
+        if (productOptional.isPresent()) {
+            Product product = productOptional.get();
+            productRepository.delete(product);
+        }
         return true;
+    }
+
+    private ProductDto toDto(Product product) {
+        return ProductDto.builder()
+                .id(product.getId())
+                .menuId(product.getMenuId())
+                .name(
+                        new TextModel(
+                                product.getNameUz(),
+                                product.getNameRu(),
+                                product.getNameEng()
+                        )
+                )
+                .image(product.getImage())
+                .costPrice(product.getCostPrice())
+                .description(product.getDescription())
+                .build();
+    }
+
+    private Product fromDto(ProductDto productDto) {
+        return Product.builder()
+                .id(productDto.getId())
+                .menuId(productDto.getMenuId())
+                .nameUz(productDto.getName().getUz())
+                .nameRu(productDto.getName().getRu())
+                .nameEng(productDto.getName().getEng())
+                .image(productDto.getImage())
+                .costPrice(productDto.getCostPrice())
+                .description(productDto.getDescription())
+                .build();
     }
 }
